Fixed a stack corruption issue
It can happen that the stack frame we want to switch to is still active,
in a rare condition: A tc_thread reads a few byte from an fd, and calls
tc_wait_fd(), although there is still input available. That call enables
the epoll-event again. Before that tc_thread makes it to the scheduler
(and epoll_wait), an other worker returns from its epoll_wait call and
switches to the same tc_thread again. That causes bad stack corruption
of course.
To circumvent that we spin in switch_to() here until the tc_thread is no
longer running