Fixed a race between removing of FDs and delivery of events on those
If an event gets delivered for a FD that was just removed
before on an other thread, we might get many weired crashes...

Solution:
  synchronize_world()

Is a function that wakes all threads, waits untill all of them reached
a common point, and contiues execution of everything after that.
By that we can be sure that no event of the-to-bee-freed FD is
hiding somwhere..

The _tc_fd_unregister(... , int sync) has the new "sync" parameter.
Call it with sync=0 if you are sure that there will be no event
on that FD. Call it with sync=1 if there might come events in.