Fixed tc_waitq_finish_wait()
tc_waitq_finish_wait() may not remove the event from any list blindly. It
only wants to remove the event from the list, where the
tc_waitq_prepare_to_wait() had put it. This caused program stalls if the
event was wrongly removed from the tc_thread's pending list.

If the event we waited for and an interrupting event (signal) come in
concurrently, re-queue the signal for later delivery and return
that the event happened.