Don't use the sync_fd anymore, but notify pthreads as needed via pgkill.
The sync_fd couldn't be directed at specific threads - it woke up
all of them, until it was read() again.
This caused a fair bit of CPU usage, especially if a pthread couldn't
immediately run because its CPU is busy elsewhere.

Furthermore there was a race in worker_after_sleep, line 1390:
When the read() from sync_fd was delayed by eg. scheduling, another
thread might have got a count of 2 from the read, so that the
other thread got blocked.
A possible workaround is to write() if c>1, or perhaps to extend
the hold time of sched.sync_lock.