Remember the stack order of events for correct wakeup.
When a tc_thread does a tc_waitq_wait_event(), and the condition uses
a libtcr event (eg. via tc_mutex_lock()), then the waitq event caused
the tc_mutex_lock():tc_waitq_finish_wait() to see a different signal,
and (erroneously) return RV_INTR.

So we store the stack of events, and if the wakeup is for a "hidden"
event we just store it on the tc->pending queue; in tc_scheduler we
look for the topmost event in the pending queue and prioritize this
one.
Signal events must not get on the event stack, of course.

Conflicts:

	tcr/threaded_cr.c