Move the waitqueue away from the stack.
Another thread might see this wq, and try to use it later
(when the spinlock is released) - but by then the wq on
the stack might be dead and overwritten.

To avoid having to use RCU or reference counting everywhere
the structure gets used in the per-tc_thread structure; that
should be safe, as on tc_die() all active operations are awaited
before quitting.