Always ping other threads to look at the immediate queue.
With 3 threads there's a race where tc_mutex_unlock() fails to
get the new thread woken up; so the lock seems to stay locked,
and everything grinds to a halt.
Unconditionally telling _some_ thread to look if there's something
that needs doing might cost a few CPU cycles, but should be safe.

That means that the distinction _iwi_immediate() and iwi_immediate
is no longer necessary, so we can drop one.